refactor(search): split /search/page.tsx into _components
Move EmptyState, SectionHeading, SearchResultCourses, and
SearchResultLessons into the page-private _components folder, and share
the privacy-aware authorLabel helper between them.
