+++ src/app/(protected)/search/_components/authorLabel.ts
@@ -1,0 +1,4 @@
+// Privacy: never derive a handle from email; keep it to the display name.
+export function authorLabel(name: string | null | undefined): string {
+  return name ?? "Anonymous";
+}
